+++ src/app/events/messageReactionAdd.ts
@@ -1,5 +1,5 @@
 import { MessageReaction, PartialUser, User } from 'discord.js';
-import handleBountyBoardChannel from './bounty/handleBountyBoardChannel';
+import messageReactionAddBounty from './bounty/messageReactionAddBounty';
 
 module.exports = {
 	name: 'messageReactionAdd',
@@ -27,10 +27,9 @@ module.exports = {
 		}
 		
 		if (user.bot) {
-			console.log('ignoring bot reactions');
 			return;
 		}
 		
-		await handleBountyBoardChannel(reaction, user as User);
+		await messageReactionAddBounty(reaction, user as User);
 	},
 };