+++ src/app/service/bounty/ClaimBounty.ts
@@ -3,6 +3,8 @@ import mongo, { Db, UpdateWriteOpResult } from 'mongodb';
 import BountyUtils from '../../utils/BountyUtils';
 import dbInstance from '../../utils/db';
 import { GuildMember, Message, MessageEmbed } from 'discord.js';
+import envUrls from '../constants/envUrls';
+import { BountyCollection } from '../../types/bounty/BountyCollection';
 
 export default async (guildMember: GuildMember, bountyId: string): Promise<any> => {
 	await BountyUtils.validateBountyId(guildMember, bountyId);
@@ -15,21 +17,20 @@ export const claimBountyForValidId = async (guildMember: GuildMember,
 	const db: Db = await dbInstance.dbConnect(constants.DB_NAME_BOUNTY_BOARD);
 	const dbCollection = db.collection(constants.DB_COLLECTION_BOUNTIES);
 	
-	const dbBountyResult = await dbCollection.findOne({
+	const dbBountyResult: BountyCollection = await dbCollection.findOne({
 		_id: new mongo.ObjectId(bountyId),
-		status: 'Open',
 	});
 	
 	await BountyUtils.checkBountyExists(guildMember, dbBountyResult, bountyId);
 
 	if (dbBountyResult.claimedBy && dbBountyResult.status != 'Open') {
 		console.log(`${bountyId} bounty not open and is claimed by ${dbBountyResult.claimedBy.discordHandle}`);
-		return guildMember.send(`Sorry <@${guildMember.user.id}>, bounty \`${bountyId}\` is not open.`);
+		return guildMember.send(`<@${guildMember.user.id}> Sorry bounty is not open. ${envUrls.BOUNTY_BOARD_URL}${bountyId}`);
 	}
 
 	if (dbBountyResult.status != 'Open') {
 		console.log(`${bountyId} bounty is not open`);
-		return guildMember.send(`Sorry bounty \`${bountyId}\` is not Open.`);
+		return guildMember.send(`<@${guildMember.user.id}> Sorry bounty is not Open. ${envUrls.BOUNTY_BOARD_URL}${bountyId}`);
 	}
 
 	const currentDate = (new Date()).toISOString();
@@ -58,20 +59,25 @@ export const claimBountyForValidId = async (guildMember: GuildMember,
 	console.log(`${bountyId} bounty claimed by ${guildMember.user.tag}`);
 	await claimBountyMessage(guildMember, dbBountyResult.discordMessageId, message);
 	
-	return guildMember.send(`<@${guildMember.user.id}> Bounty claimed! Feel free to reach out at any time ${constants.BOUNTY_BOARD_URL}${bountyId}`);
+	return guildMember.send(`<@${guildMember.user.id}> Bounty claimed! Feel free to reach out at any time ${envUrls.BOUNTY_BOARD_URL}${bountyId}`);
 };
 
 export const claimBountyMessage = async (guildMember: GuildMember, bountyMessageId: string, message?: Message): Promise<any> => {
-	message = (message === null) ? await BountyUtils.getBountyMessage(guildMember, bountyMessageId) : message;
+	console.log(`attempting to claim bountyMessageId: ${bountyMessageId}`);
+	message = await BountyUtils.getBountyMessage(guildMember, bountyMessageId, message);
 	
 	const embedMessage: MessageEmbed = message.embeds[0];
 	embedMessage.fields[1].value = 'In-Progress';
 	embedMessage.setColor('#d39e00');
 	embedMessage.addField('Claimed By', guildMember.user.tag, true);
-	embedMessage.setFooter('📮 - submit | 🆘 - help');
+	embedMessage.setFooter('📮 - submit | 🔄 - refresh | 🆘 - help');
 	await message.edit(embedMessage);
+	addClaimReactions(message);
+};
 
-	await message.reactions.removeAll();
-	await message.react('📮');
-	await message.react('🆘');
+export const addClaimReactions = (message: Message): void => {
+	message.reactions.removeAll();
+	message.react('📮');
+	message.react('🔄');
+	message.react('🆘');
 };