+++ src/app/service/bounty/CompleteBounty.ts
@@ -3,13 +3,16 @@ import BountyUtils from '../../utils/BountyUtils';
 import mongo, { Db, UpdateWriteOpResult } from 'mongodb';
 import dbInstance from '../../utils/db';
 import constants from '../constants/constants';
+import { seekHelpValidBountyId } from './SeekHelpBounty';
+import { BountyCollection } from '../../types/bounty/BountyCollection';
 
 export default async (guildMember: GuildMember, bountyId: string, isComplete: boolean): Promise<any> => {
-	if (!isComplete) {
-		// TODO: execute SOS help
-		return;
-	}
 	await BountyUtils.validateBountyId(guildMember, bountyId);
+
+	if (!isComplete) {
+		return seekHelpValidBountyId(guildMember, bountyId);
+	}
+	
 	return completeBountyForValidId(guildMember, bountyId);
 };
 
@@ -19,7 +22,7 @@ export const completeBountyForValidId = async (guildMember: GuildMember,
 	const db: Db = await dbInstance.dbConnect(constants.DB_NAME_BOUNTY_BOARD);
 	const dbCollection = db.collection(constants.DB_COLLECTION_BOUNTIES);
 
-	const dbBountyResult = await dbCollection.findOne({
+	const dbBountyResult: BountyCollection = await dbCollection.findOne({
 		_id: new mongo.ObjectId(bountyId),
 		status: 'In-Review',
 	});
@@ -63,19 +66,22 @@ export const completeBountyForValidId = async (guildMember: GuildMember,
 	console.log(`${bountyId} bounty reviewed by ${guildMember.user.tag}`);
 	await completeBountyMessage(guildMember, dbBountyResult.discordMessageId, message);
 
-	return guildMember.send(`<@${guildMember.user.id}> Bounty in review! Look out for a follow up message from <@${dbBountyResult.createdBy.id}>`);
+	return guildMember.send(`<@${guildMember.user.id}> Bounty in review! Look out for a follow up message from <@${dbBountyResult.createdBy.discordId}>`);
 };
 
 export const completeBountyMessage = async (guildMember: GuildMember, bountyMessageId: string, message?: Message): Promise<any> => {
-	message = (message === null) ? await BountyUtils.getBountyMessage(guildMember, bountyMessageId) : message;
+	message = await BountyUtils.getBountyMessage(guildMember, bountyMessageId, message);
 
 	const embedMessage: MessageEmbed = message.embeds[0];
 	embedMessage.fields[1].value = 'Completed';
 	embedMessage.setColor('#1d2124');
 	embedMessage.addField('Reviewed By', guildMember.user.tag, true);
-	embedMessage.setFooter('🆘 - help | bounty complete');
+	embedMessage.setFooter('🆘 - help');
 	await message.edit(embedMessage);
+	addCompletedReactions(message);
+};
 
-	await message.reactions.removeAll();
-	await message.react('🆘');
+export const addCompletedReactions = (message: Message): void => {
+	message.reactions.removeAll();
+	message.react('🆘');
 };