+++ src/app/service/bounty/create/PublishBounty.ts
@@ -5,6 +5,8 @@ import { GuildMember, Message, MessageOptions, TextChannel } from 'discord.js';
 import dbInstance from '../../../utils/db';
 import channelIDs from '../../constants/channelIDs';
 import ServiceUtils from '../../../utils/ServiceUtils';
+import envUrls from '../../constants/envUrls';
+import { BountyCollection } from '../../../types/bounty/BountyCollection';
 
 export default async (guildMember: GuildMember, bountyId: string): Promise<any> => {
 	await BountyUtils.validateBountyId(guildMember, bountyId);
@@ -16,7 +18,7 @@ export const finalizeBounty = async (guildMember: GuildMember, bountyId: string)
 
 	const db: Db = await dbInstance.dbConnect(constants.DB_NAME_BOUNTY_BOARD);
 	const dbCollection = db.collection(constants.DB_COLLECTION_BOUNTIES);
-	const dbBountyResult = await dbCollection.findOne({
+	const dbBountyResult: BountyCollection = await dbCollection.findOne({
 		_id: new mongo.ObjectId(bountyId),
 		status: 'Draft',
 	});
@@ -27,37 +29,12 @@ export const finalizeBounty = async (guildMember: GuildMember, bountyId: string)
 		console.log(`${bountyId} bounty is not drafted`);
 		return guildMember.send(`<@${guildMember.user.id}> Sorry bounty is not drafted.`);
 	}
-	const messageOptions: MessageOptions = {
-		embed: {
-			color: '#1e7e34',
-			title: dbBountyResult.title,
-			url: constants.BOUNTY_BOARD_URL + dbBountyResult._id,
-			author: {
-				icon_url: guildMember.user.avatarURL(),
-				name: dbBountyResult.createdBy.discordHandle,
-			},
-			description: dbBountyResult.summary,
-			fields: [
-				{ name: 'Reward', value: dbBountyResult.reward.amount + ' ' + dbBountyResult.reward.currency, inline: true },
-				{ name: 'Status', value: 'Open', inline: true },
-				{ name: 'Deadline', value: ServiceUtils.formatDisplayDate(dbBountyResult.dueAt), inline: true },
-				{ name: 'Criteria', value: dbBountyResult.criteria },
-				{ name: 'Summary', value: dbBountyResult.description },
-				{ name: 'HashId', value: dbBountyResult._id },
-				{ name: 'Created By', value: dbBountyResult.createdBy.discordHandle, inline: true },
-			],
-			timestamp: new Date(),
-			footer: {
-				text: '🏴 - start | 📝 - edit | ❌ - delete',
-			},
-		},
-	};
+	const messageOptions: MessageOptions = generateEmbedMessage(dbBountyResult, guildMember.user.avatarURL());
 
 	const bountyChannel: TextChannel = guildMember.guild.channels.cache.get(channelIDs.bountyBoard) as TextChannel;
 	const bountyMessage: Message = await bountyChannel.send(messageOptions) as Message;
-	await bountyMessage.react('🏴');
-	await bountyMessage.react('📝');
-	await bountyMessage.react('❌');
+	console.log('bounty published to #bounty-board');
+	addPublishReactions(bountyMessage);
 
 	const currentDate = (new Date()).toISOString();
 	const writeResult: UpdateWriteOpResult = await dbCollection.updateOne(dbBountyResult, {
@@ -80,5 +57,40 @@ export const finalizeBounty = async (guildMember: GuildMember, bountyId: string)
 
 	await dbInstance.close();
 
-	return guildMember.send(`<@${guildMember.user.id}> Bounty published to #🧀-bounty-board and the website! ${constants.BOUNTY_BOARD_URL}${bountyId}`);
+	return guildMember.send(`<@${guildMember.user.id}> Bounty published to #🧀-bounty-board and the website! ${envUrls.BOUNTY_BOARD_URL}${bountyId}`);
+};
+
+export const addPublishReactions = (message: Message): void => {
+	message.reactions.removeAll();
+	message.react('🏴');
+	message.react('🔄');
+	message.react('📝');
+	message.react('❌');
+};
+
+export const generateEmbedMessage = (dbBounty: BountyCollection, iconUrl?: string): MessageOptions => {
+	return {
+		embed: {
+			color: '#1e7e34',
+			title: dbBounty.title,
+			url: envUrls.BOUNTY_BOARD_URL + dbBounty._id,
+			author: {
+				icon_url: iconUrl,
+				name: dbBounty.createdBy.discordHandle,
+			},
+			description: dbBounty.description,
+			fields: [
+				{ name: 'Reward', value: dbBounty.reward.amount + ' ' + dbBounty.reward.currency, inline: true },
+				{ name: 'Status', value: 'Open', inline: true },
+				{ name: 'Deadline', value: ServiceUtils.formatDisplayDate(dbBounty.dueAt), inline: true },
+				{ name: 'Criteria', value: dbBounty.criteria },
+				{ name: 'HashId', value: dbBounty._id },
+				{ name: 'Created By', value: dbBounty.createdBy.discordHandle, inline: true },
+			],
+			timestamp: new Date(),
+			footer: {
+				text: '🏴 - start | 🔄 - refresh | 📝 - edit | ❌ - delete',
+			},
+		},
+	};
 };