+++ src/app/service/constants/channelIDs.ts
@@ -1,3 +1,3 @@
-export default {
+export default Object.freeze({
 	bountyBoard: process.env.DISCORD_CHANNEL_BOUNTY_BOARD_ID,
-};
+});