+++ src/app/types/bounty/BountyReward.ts
@@ -1,4 +1,5 @@
 export type BountyReward = {
 	amount: number,
-	currencySymbol: string
+	currencySymbol: string,
+	scale: number
 };